requirements.txt
SQLAlchemy==2.0.36
fastapi[standard]==0.115.6
pydantic==2.10.3
pytest==8.3.4
passlib==1.7.4
requests==2.32.3
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
uvicorn~=0.34.0
python-dotenv~=1.0.1